# backend/API_DOCUMENTATION.yaml
openapi: 3.0.0
info:
  title: LocateMe API
  description: Real-time location tracking system with secure authentication and OTP verification
  version: 1.0.0
  contact:
    name: LocateMe Team
    email: support@locateme.com
  license:
    name: ISC
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://api.locateme.com/api
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained after login/registration
  schemas:
...
    Location:
      type: object
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        address:
          type: string
        method:
          type: string
          example: "geolocation"
        timestamp:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
          description: Only in development mode
        details:
          type: string
          description: Only in development mode
paths:
...
  /auth/share-location:
    post:
      summary: Share current location with a tracking requester
      operationId: shareLocationWithRequester
      tags:
        - Location
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                requesterEmail:
                  type: string
                  format: email
                latitude:
                  type: number
                  format: double
                  minimum: -90
                  maximum: 90
                longitude:
                  type: number
                  format: double
                  minimum: -180
                  maximum: 180
                address:
                  type: string
                method:
                  type: string
                  default: "geolocation"
              required:
                - requesterEmail
                - latitude
                - longitude
              example:
                requesterEmail: "john@example.com"
                latitude: 40.7128
                longitude: -74.0060
                address: "New York, USA"
                method: "geolocation"
      responses:
        "200":
          description: Location shared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  locationData:
                    $ref: "#/components/schemas/Location"
        "400":
          description: Invalid location data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: No active tracking request from this requester
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/tracked-user-location:
    get:
      summary: Get latest location of a tracked user
      operationId: getTrackedUserLocation
      tags:
        - Location
      security:
        - BearerAuth: []
      parameters:
        - name: targetEmail
          in: query
          required: true
          schema:
            type: string
            format: email
          example: "jane@example.com"
      responses:
        "200":
          description: Latest location data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  targetEmail:
                    type: string
                  lastLocation:
                    $ref: "#/components/schemas/Location"
                  locationHistory:
                    type: array
                    items:
                      $ref: "#/components/schemas/Location"
                    description: Last 10 location updates
                  updatedAt:
                    type: string
                    format: date-time
        "400":
          description: Missing target email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: No active tracking permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...